Revert "Dynamic getCorrections Timeout"
This reverts commit ee29e3448cef0778bbf5e7ffbcd165df1b309850.

+++ Passions/PASResources.h
@@ -14,7 +14,7 @@ extern NSString * const kPASParseClientKey;
 extern NSString * const kPASParseMasterKey;
 
 extern NSString * const kPASLastFmApiKey;
-extern float const kPASLastFmTimeoutInSec;
+extern CGFloat const kPASLastFmTimeoutInSec;
 
 extern NSString * const kPASSpotifyClientId;
